Create ReverseFactorial
Here I made a method to take a number and keep dividing it by the number of iterations its run through. This can return the factorial value of a number.  I want to add a way to test the input and see if the number actually has a factorial value, because even though the method works, if called on a number that is not a factorial, it will run anyway.  Does anyone have any suggestions?